requirements.txt
ollama
nltk
sentence-transformers
pymongo
tk
requests
beautifulsoup4
python-docx
googlesearch-python
certifi